Refactor main_utils to utils
