+++ django-caretaker/caretaker/management/commands/install_cron.py
@@ -4,7 +4,7 @@ from crontab import CronTab
 from django.conf import settings
 from django.core.management.base import BaseCommand
 
-from caretaker.main_utils import log
+from caretaker.utils import log
 
 
 def find_job(tab, comment):
