+++ django-caretaker/caretaker/management/commands/list_backups.py
@@ -3,7 +3,7 @@ from django.conf import settings
 from django.core.management.base import BaseCommand
 
 from caretaker.backend.abstract_backend import BackendFactory
-from caretaker.main_utils import log
+from caretaker.utils import log
 
 
 class Command(BaseCommand):
