+++ django-caretaker/caretaker/management/commands/push_backup.py
@@ -4,7 +4,7 @@ from django.conf import settings
 from django.core.management.base import BaseCommand
 
 from caretaker.backend.abstract_backend import BackendFactory, StoreOutcome
-from caretaker.main_utils import log
+from caretaker.utils import log
 
 
 class Command(BaseCommand):
