+++ django-caretaker/caretaker/management/commands/run_backup.py
@@ -4,7 +4,7 @@ from django.conf import settings
 from django.core.management.base import BaseCommand
 
 from caretaker.backend.abstract_backend import BackendFactory
-from caretaker.main_utils import log
+from caretaker.utils import log
 from caretaker.management.commands.create_backup import Command as CreateCommand
 from caretaker.management.commands.push_backup import Command as PushCommand
