+++ django-caretaker/caretaker/utils/file.py
@@ -1,0 +1,5 @@
+from pathlib import Path
+
+
+def normalize_path(path: str) -> Path:
+    return Path(path).expanduser()
