+++ django-caretaker/caretaker/utils/zip.py
